requirements.txt
arrow==1.3.0
asgiref==3.7.2
binaryornot==0.4.4
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
cookiecutter==2.4.0
cryptography==41.0.4
deskew==1.4.3
Django==4.2.6
django-appconf==1.0.5
google-api-core==2.12.0
google-auth==2.23.2
google-cloud-vision==3.4.4
googleapis-common-protos==1.60.0
grpcio==1.59.0
grpcio-status==1.59.0
idna==3.4
imageio==2.31.5
Jinja2==3.1.2
lazy_loader==0.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
networkx==3.1
numpy==1.26.0
opencv-python==4.8.1.78
packaging==23.2
pdfminer.six==20221105
pdfplumber==0.10.2
Pillow==10.0.1
proto-plus==1.22.3
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.16.1
pypdfium2==4.20.0
python-dateutil==2.8.2
python-decouple==3.8
python-slugify==8.0.1
PyYAML==6.0.1
rcssmin==1.1.1
requests==2.31.0
rich==13.6.0
rjsmin==1.2.1
rsa==4.9
scikit-image==0.22.0
scipy==1.11.3
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
tifffile==2023.9.26
types-python-dateutil==2.8.19.14
tzdata==2023.3
urllib3==2.0.6
